refactor(models): breakdown models into separated rust modules
